requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distro==1.9.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openai==1.53.0
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pip==24.3.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypdfium2==4.30.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
